pyproject.toml
[tool.poetry]

name = "instill-python-sdk"
version = "0.1.0"
description = "python sdk for Instill AI products"

packages = [{ include = "instill_sdk" }]

license = "MIT"
authors = ["instill-ai <hello@instill.tech>"]

readme = "README.md"
homepage = "https://pypi.org/project/instill-python-sdk"
documentation = "https://instill-python-sdk.readthedocs.io"
repository = "https://github.com/instill-ai/python-sdk"

keywords = [
]
classifiers = [
    # TODO: update this list to match your application: https://pypi.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
]

[tool.poetry.dependencies]

python = "^3.8"

# TODO: Remove these and add your library's requirements
click = "*"
minilog = "*"
grpcio = "1.48.2"
protobuf = "^4.24.2"
google-api-core = "^2.11.1"
googleapis-common-protos = "^1.60.0"
protoc-gen-openapiv2 = "^0.0.1"
types-protobuf = "^4.24.0.1"

[tool.poetry.dev-dependencies]

# Formatters
black = "^22.1"
tomli = "*" # missing 'black' dependency
isort = "^5.10"

# Linters
mypy = "^1.0"
pydocstyle = "^6.1"
pylint = "~2.15"
wrapt = "*" # missing 'pylint' dependency

# Testing
pytest = "^7.3"
pytest-describe = "^2.0"
pytest-expecter = "^3.0"
pytest-random = "*"
pytest-cov = "^4.1"
freezegun = "*"

# Reports
coveragespace = "^6.0"

# Documentation
mkdocs = "~1.3"
pygments = "^2.11.1"

# Tooling
pyinstaller = "*"
sniffer = "*"
MacFSEvents = { version = "*", platform = "darwin" }
pync = { version = "*", platform = "darwin" }
ipython = "^7.12.0"

[tool.poetry.scripts]

# python-sdk = "instill_sdk.cli:main"

[tool.black]

quiet = true
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | protogen
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]

profile = "black"
skip_glob = ["**/protogen/*"]

[tool.mypy]

ignore_missing_imports = true
no_implicit_optional = true
check_untyped_defs = true

cache_dir = ".cache/mypy/"

[tool.pytest.ini_options]

addopts = """
--strict-markers

-r sxX
--show-capture=log

--cov-report=html
--cov-report=term-missing:skip-covered
--no-cov-on-fail
"""

cache_dir = ".cache/pytest/"

markers = []

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
